first stage refactor
